requirements.txt
Flask
Json
mysql.connector